rewrite the timer context store
